Refactor home page components and enhance UI sections
- Replaced old section components with updated versions: HeroSection, WhyMarketingSimplifiedSection, SmartWorkflowSection, FeatureCarousel, HowItWorksSection, TestimonialsSection, IntegrationsSection, CtaSection, and FooterSection.
- Removed the AutomatedEfficiencySection and InsightfulAnalyticsSection for a cleaner structure.
- Updated the AiWorkflowsSection to clarify workflow triggers.
- Improved the CtaSection with new animations and layout adjustments for better user engagement.

This commit streamlines the home page layout and enhances the overall user experience with modernized components.
