+++ frontend/src/components/home/CtaSection.tsx
@@ -1,63 +1,66 @@
-import React from 'react';
-import { ArrowRight } from 'lucide-react';
+"use client"
+
+import { useRef } from "react"
+import { motion, useInView } from "framer-motion"
+import { ArrowRight } from "lucide-react"
+import { Button } from "@/components/ui/button"
 
 type CtaSectionProps = {
-  onGetStartedClick: () => void;
-  onRedirectToLogin: () => void;
-};
+  onGetStartedClick: () => void
+}
 
-export default function CtaSection({ onGetStartedClick, onRedirectToLogin }: CtaSectionProps) {
+export default function CtaSection({ onGetStartedClick }: CtaSectionProps) {
+  const ref = useRef(null)
+  const isInView = useInView(ref, { once: true, margin: "-100px" })
+
   return (
-    <>
-      {/* CTA Section - Desktop */}
-      <section className="hidden md:block py-32 px-6 bg-blue-50">
-        <div className="max-w-4xl mx-auto text-center">
-          <h2 className="text-4xl md:text-5xl font-bold text-gray-900 mb-6 leading-[1.8]">
-            Ready to transform your<br />
-            Ad operations?
-          </h2>
-          <p className="text-xl text-gray-600 mb-10 max-w-2xl mx-auto whitespace-nowrap">
-            Simplify every stage of your advertising lifecycle — from planning to performance.
-          </p>
-          <div className="flex flex-col sm:flex-row gap-8 justify-center items-center">
-            <button
-              onClick={onGetStartedClick}
-              className="px-8 py-4 bg-blue-800 text-white rounded-full hover:bg-blue-900 transition font-medium text-lg inline-flex items-center gap-2 shadow-lg justify-center w-[200px]"
-            >
-              Get Started
-              <ArrowRight className="w-5 h-5" />
-            </button>
-            <button
-              onClick={onRedirectToLogin}
-              className="px-8 py-4 border-2 border-blue-600 text-blue-600 bg-white rounded-full hover:bg-blue-50 transition font-medium text-lg shadow-lg inline-flex items-center justify-center w-[200px]"
-            >
-              Contact Us
-            </button>
+    <section className="py-24" ref={ref}>
+      <div className="max-w-7xl mx-auto px-4">
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          animate={isInView ? { opacity: 1, y: 0 } : {}}
+          transition={{ duration: 0.6 }}
+          className="relative overflow-hidden rounded-3xl bg-brand-teal/5 border border-brand-teal/15 p-12 lg:p-20 text-center"
+        >
+          <div className="absolute inset-0 pointer-events-none">
+            <div className="absolute top-0 left-0 w-96 h-96 bg-brand-teal/10 rounded-full blur-[80px] -translate-x-1/2 -translate-y-1/2" />
+            <div className="absolute bottom-0 right-0 w-96 h-96 bg-brand-lime/10 rounded-full blur-[80px] translate-x-1/2 translate-y-1/2" />
           </div>
-        </div>
-      </section>
 
-      {/* CTA Section - Mobile */}
-      <section className="block md:hidden py-10 px-6 bg-white">
-        <div className="max-w-4xl mx-auto text-center">
-          <h2 className="text-2xl font-bold text-gray-900 mb-6 leading-tight">
-            Ready to transform your<br />
-            Ad operations?
-          </h2>
-          <p className="text-base text-gray-700 mb-10 leading-relaxed">
-            Simplify every stage of your advertising lifecycle — from planning to performance.
-          </p>
-          <div className="flex flex-col sm:flex-row gap-8 justify-center items-center">
-            <button className="items-center justify-center px-6 py-3 bg-blue-800 text-white rounded-full hover:bg-blue-900 transition text-center font-medium text-sm inline-flex gap-2 shadow-lg w-[60%]">
-              See All Reviews
-              <ArrowRight className="w-4 h-4" />
-            </button>
-            <button className="px-6 py-3 border-2 border-blue-600 text-blue-600 bg-white rounded-full hover:bg-blue-50 transition font-medium text-sm shadow-lg inline-flex items-center justify-center w-[60%]">
-              Contact Us
-            </button>
+          <div className="relative z-10 max-w-4xl mx-auto space-y-6">
+            <h2 className="text-4xl lg:text-5xl font-bold text-slate-900 leading-tight">
+              Ready to transform your
+              <br />
+              <span className="bg-gradient-to-r from-brand-teal to-brand-lime bg-clip-text text-transparent">
+                Ad operations
+              </span>
+              ?
+            </h2>
+
+            <p className="text-xl text-gray-600 max-w-2xl mx-auto">
+              Simplify every stage of your advertising lifecycle - from planning to performance.
+            </p>
+
+            <div className="flex flex-col sm:flex-row gap-4 justify-center pt-4">
+              <Button
+                size="lg"
+                onClick={onGetStartedClick}
+                className="text-lg px-8 h-14 rounded-full bg-brand-gradient hover:saturate-150 transition-all text-white glow-brand"
+              >
+                Get Started
+                <ArrowRight className="ml-2 w-5 h-5" />
+              </Button>
+              <Button
+                size="lg"
+                variant="outline"
+                className="text-lg px-8 h-14 rounded-full border-2 border-brand-teal/40 text-brand-teal bg-white hover:bg-brand-teal/5"
+              >
+                Contact Us
+              </Button>
+            </div>
           </div>
-        </div>
-      </section>
-    </>
-  );
+        </motion.div>
+      </div>
+    </section>
+  )
 }
