+++ frontend/src/components/home/IntegrationsSection.tsx
@@ -22,7 +22,7 @@ const stats = [
   { value: "24/7", label: "Support" },
 ]
 
-export default function NewIntegrationsSection() {
+export default function IntegrationsSection() {
   const ref = useRef(null)
   const isInView = useInView(ref, { once: true, margin: "-100px" })
