+++ frontend/src/components/home/TestimonialsSection.tsx
@@ -1,368 +1,165 @@
-import React from 'react';
-import { ArrowRight } from 'lucide-react';
+"use client"
 
-type TestimonialsSectionProps = {
-  onRedirectToLogin: () => void;
-  onGetStartedClick: () => void;
-};
-export default function TestimonialsSection({ onRedirectToLogin, onGetStartedClick }: TestimonialsSectionProps) {
+import { useRef } from "react"
+import { motion, useInView } from "framer-motion"
+import { ArrowRight, Quote, Star } from "lucide-react"
+import { Avatar } from "@/components/avatar/Avatar"
+
+const testimonials = [
+  {
+    id: 1,
+    quote: "Finally, one place to manage budgets, assets, and approvals. We cut our setup time by half.",
+    author: "Jack Wilson",
+    role: "Senior Media Buyer",
+    company: "Agency Partners",
+    avatar: "/avatars/person-1.jpg",
+    rating: 5,
+  },
+  {
+    id: 2,
+    quote: "I love how the review and approval flow feels natural. No more endless Slack threads.",
+    author: "Eve Turner",
+    role: "Designer",
+    company: "Creative Studio",
+    avatar: "/avatars/person-2.jpg",
+    rating: 5,
+  },
+  {
+    id: 3,
+    quote: "The reporting dashboard gives instant clarity. We spot underperforming channels in minutes.",
+    author: "Grace Lee",
+    role: "Data Analyst",
+    company: "Growth Labs",
+    avatar: "/avatars/person-3.jpg",
+    rating: 5,
+  },
+  {
+    id: 4,
+    quote: "The AI Agent has transformed how we manage campaigns. It catches issues before they become problems.",
+    author: "Marcus Chen",
+    role: "Marketing Director",
+    company: "Tech Ventures",
+    avatar: "/avatars/person-4.jpg",
+    rating: 5,
+  },
+  {
+    id: 5,
+    quote: "Cross-platform campaign management used to take hours. Now it takes minutes.",
+    author: "Sarah Martinez",
+    role: "Media Specialist",
+    company: "Digital First",
+    avatar: "/avatars/person-9.jpg",
+    rating: 5,
+  },
+  {
+    id: 6,
+    quote: "Our team collaboration improved dramatically. Everyone knows exactly what's happening.",
+    author: "Tom Anderson",
+    role: "Project Manager",
+    company: "Brand Works",
+    avatar: "/avatars/person-6.jpg",
+    rating: 5,
+  },
+]
+
+function TestimonialCard({ testimonial }: { testimonial: (typeof testimonials)[0] }) {
   return (
-    <>
-      {/* Testimonials Section -Desktop */}
-      <section id="testimonials" className="hidden md:block py-20 px-6 bg-blue-50 rounded-t-3xl">
-        <div className="max-w-7xl mx-auto">
-          <div className="text-center mb-16">
-            <h2 className="text-4xl md:text-5xl font-bold text-gray-900 mb-4">
-              What teams love about<br />
-              Marketing Simplified?
-            </h2>
-            <p className="text-lg text-gray-600 mx-auto text-center">
-              From setup to reporting, Marketing Simplified helps teams stay aligned, efficient, and confident in every campaign.
-            </p>
-          </div>
+    <div className="glass-card rounded-2xl p-4 sm:p-6 min-w-[260px] max-w-[calc(100vw-3rem)] sm:min-w-[350px] sm:max-w-none flex-shrink-0 hover:shadow-lg transition-all duration-300 hover:-translate-y-1 cursor-pointer flex flex-col h-auto min-h-[200px] sm:h-[280px]">
+      <div className="mb-2 sm:mb-4">
+        <Quote className="w-6 h-6 sm:w-8 sm:h-8 text-brand-teal/30" />
+      </div>
 
-          {/* Testimonial Cards with Carousel */}
-          <div className="mb-12 relative overflow-hidden">
-            <div className="flex gap-6 animate-marquee group hover:[animation-play-state:paused]">
-              {/* Testimonial 1 */}
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  Finally, one place to manage<br />
-                  budgets, assets, and approvals.<br />
-                  We cut our setup time by half.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-blue-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=68"
-                      alt="Jack Wilson"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Jack Wilson</div>
-                    <div className="text-sm text-gray-600">Senior Media Buyer</div>
-                  </div>
-                </div>
+      <div className="flex gap-1 mb-2 sm:mb-3">
+        {[...Array(testimonial.rating)].map((_, i) => (
+          <Star key={i} className="w-4 h-4 text-amber-400 fill-amber-400" />
+        ))}
+      </div>
+
+      <p className="text-sm sm:text-base text-gray-700 mb-4 sm:mb-6 flex-grow leading-relaxed">
+        &quot;{testimonial.quote}&quot;
+      </p>
+
+      <div className="flex items-center gap-3 mt-auto pt-4 border-t border-gray-100">
+        <Avatar
+          src={testimonial.avatar}
+          alt={testimonial.author}
+          size="lg"
+          fallback={testimonial.author.split(" ").map((n) => n[0]).join("")}
+          className="ring-2 ring-white shadow flex-shrink-0"
+        />
+        <div>
+          <div className="font-bold text-gray-900">{testimonial.author}</div>
+          <div className="text-sm text-gray-600">{testimonial.role}</div>
+          <div className="text-xs text-gray-400">{testimonial.company}</div>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default function TestimonialsSection() {
+  const ref = useRef(null)
+  const isInView = useInView(ref, { once: true, margin: "-100px" })
+
+  return (
+    <section ref={ref} className="py-20 bg-gray-50/50 overflow-hidden">
+      <div className="max-w-7xl mx-auto px-6">
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          animate={isInView ? { opacity: 1, y: 0 } : {}}
+          transition={{ duration: 0.6 }}
+          className="text-center mb-12"
+        >
+          <h2 className="text-4xl lg:text-5xl font-bold text-slate-900 mb-4">
+            What teams{" "}
+            <span className="bg-gradient-to-r from-brand-teal to-brand-lime bg-clip-text text-transparent">
+              love
+            </span>{" "}
+            about us
+          </h2>
+          <p className="text-xl text-gray-600 max-w-2xl mx-auto">
+            From setup to reporting, Marketing Simplified helps teams stay aligned, efficient, and confident in every
+            campaign.
+          </p>
+        </motion.div>
+      </div>
+
+      <motion.div
+        initial={{ opacity: 0 }}
+        animate={isInView ? { opacity: 1 } : {}}
+        transition={{ duration: 0.6, delay: 0.2 }}
+        className="relative mb-12"
+      >
+        <div className="absolute left-0 top-0 bottom-0 w-32 bg-gradient-to-r from-gray-50/80 to-transparent z-10 pointer-events-none" />
+        <div className="absolute right-0 top-0 bottom-0 w-32 bg-gradient-to-l from-gray-50/80 to-transparent z-10 pointer-events-none" />
+
+        <div className="overflow-hidden">
+          <div className="flex w-max animate-marquee hover:[animation-play-state:paused]">
+            {testimonials.map((testimonial) => (
+              <div key={testimonial.id} className="pr-6">
+                <TestimonialCard testimonial={testimonial} />
               </div>
-
-              {/* Testimonial 2 */}
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  I love how the review and<br />
-                  approval flow feels natural.<br />
-                  No more endless Slack threads.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-green-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=47"
-                      alt="Eve Turner"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Eve Turner</div>
-                    <div className="text-sm text-gray-600">Designer</div>
-                  </div>
-                </div>
+            ))}
+            {testimonials.map((testimonial) => (
+              <div key={`dup-${testimonial.id}`} className="pr-6">
+                <TestimonialCard testimonial={testimonial} />
               </div>
-
-              {/* Testimonial 3 */}
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  The reporting dashboard gives<br />
-                  instant clarity. We spot<br />
-                  underperforming channels in<br />
-                  minutes.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-orange-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=59"
-                      alt="Grace Lee"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Grace Lee</div>
-                    <div className="text-sm text-gray-600">Data Analyst</div>
-                  </div>
-                </div>
-              </div>
-
-              {/* Duplicate cards for seamless loop */}
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  Finally, one place to manage<br />
-                  budgets, assets, and approvals.<br />
-                  We cut our setup time by half.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-blue-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=68"
-                      alt="Jack Wilson"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Jack Wilson</div>
-                    <div className="text-sm text-gray-600">Senior Media Buyer</div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  I love how the review and<br />
-                  approval flow feels natural.<br />
-                  No more endless Slack threads.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-green-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=47"
-                      alt="Eve Turner"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Eve Turner</div>
-                    <div className="text-sm text-gray-600">Designer</div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="bg-white rounded-xl shadow-lg p-6 min-w-[350px] flex-shrink-0 hover:scale-110 transition-transform duration-300 cursor-pointer flex flex-col">
-                <p className="text-gray-700 mb-4 italic flex-grow relative mx-auto text-center">
-                  <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-12 w-4 h-4" />
-                  <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-12 w-4 h-4" />
-                  The reporting dashboard gives<br />
-                  instant clarity. We spot<br />
-                  underperforming channels in<br />
-                  minutes.
-                </p>
-                <div className="flex items-center gap-3 mt-auto">
-                  <div className="w-12 h-12 bg-orange-100 rounded-full overflow-hidden flex-shrink-0">
-                    <img
-                      src="https://i.pravatar.cc/150?img=59"
-                      alt="Grace Lee"
-                      className="w-full h-full object-cover"
-                    />
-                  </div>
-                  <div>
-                    <div className="font-bold text-gray-900">Grace Lee</div>
-                    <div className="text-sm text-gray-600">Data Analyst</div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
-          <div className="text-center">
-            <button
-              onClick={onRedirectToLogin}
-              className="px-8 py-4 bg-blue-800 text-white rounded-full hover:bg-blue-900 transition font-medium text-lg inline-flex items-center gap-2 shadow-lg"
-            >
-              See All Reviews
-              <ArrowRight className="w-5 h-5" />
-            </button>
+            ))}
           </div>
         </div>
-      </section>
+      </motion.div>
 
-      <section className="block md:hidden py-10 px-6 bg-white">
-        <div className="max-w-4xl mx-auto">
-          <div className="bg-blue-50 rounded-2xl p-8 mb-6">
-            {/* How it works Section */}
-            <div className="text-center">
-              <h2 className="text-2xl font-bold text-gray-900 mb-4 leading-tight">
-                How it works?
-              </h2>
-              <p className="text-base text-gray-700 mb-6 leading-relaxed">
-                From setup to optimization,<br />
-                Marketing Simplified keeps your campaigns moving effortlessly.
-              </p>
-            </div>
-
-            <div className="overflow-x-auto scrollbar-hide snap-x snap-mandatory scroll-smooth mb-2 pb-4 -mx-14 px-10">
-              <div className="flex items-center gap-4 min-w-max">
-                {/* Step 1 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 pt-0 w-60 hover:shadow-xl transition flex flex-col flex-shrink-0">
-                  <div className="flex items-center justify-center h-32">
-                    <img src="/step1.png" alt="Plan & Assign" className="w-24 h-24 object-contain" />
-                  </div>
-                  <div className="text-center flex-shrink-0">
-                    <p className="text-sm text-gray-600 mb-1">Step 1</p>
-                    <h3 className="text-md font-bold text-gray-900 mb-2">Plan & Assign</h3>
-                    <p className="text-sm text-gray-600">
-                      Define roles, permissions, and tasks in one place.
-                    </p>
-                  </div>
-                </div>
-
-                {/* Step 2 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 pt-0 w-60 hover:shadow-xl transition flex flex-col flex-shrink-0">
-                  <div className="flex items-center justify-center h-32">
-                    <img src="/step2.png" alt="Execute & Collab" className="w-24 h-24 object-contain" />
-                  </div>
-                  <div className="text-center flex-shrink-0">
-                    <p className="text-sm text-gray-600 mb-1">Step 2</p>
-                    <h3 className="text-md font-bold text-gray-900 mb-2">Execute & Collab</h3>
-                    <p className="text-sm text-gray-600">
-                      Manage assets, budgets, and channels seamlessly.
-                    </p>
-                  </div>
-                </div>
-
-                {/* Step 3 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 pt-0 w-60 hover:shadow-xl transition flex flex-col flex-shrink-0">
-                  <div className="flex items-center justify-center h-32">
-                    <img src="/step3.png" alt="Automate & Notify" className="w-[72px] h-[72px] object-contain" />
-                  </div>
-                  <div className="text-center flex-shrink-0">
-                    <p className="text-sm text-gray-600 mb-1">Step 3</p>
-                    <h3 className="text-md font-bold text-gray-900 mb-2">Automate & Notify</h3>
-                    <p className="text-sm text-gray-600">
-                      Trigger smart workflows and stay informed in real time.
-                    </p>
-                  </div>
-                </div>
-
-                {/* Step 4 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 pt-0 w-60 hover:shadow-xl transition flex flex-col flex-shrink-0">
-                  <div className="flex items-center justify-center h-32">
-                    <img src="/step4.png" alt="Analyze & Optimize" className="w-[72px] h-[72px] object-contain" />
-                  </div>
-                  <div className="text-center flex-shrink-0">
-                    <p className="text-sm text-gray-600 mb-1">Step 4</p>
-                    <h3 className="text-md font-bold text-gray-900 mb-2">Analyze & Optimize</h3>
-                    <p className="text-sm text-gray-600">
-                      Turn data into insights for your next campaign.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            <div className="text-center">
-              <button
-                onClick={onGetStartedClick}
-                className="flex items-center px-6 py-3 bg-blue-800 text-white rounded-full hover:bg-blue-900 transition text-center font-medium text-sm gap-4 shadow-lg"
-              >
-                Get Started
-                <ArrowRight className="w-4 h-4" />
-              </button>
-            </div>
-
-            {/* Testimonials Section */}
-            <div className="text-center pt-6">
-              <h2 className="text-2xl font-bold text-gray-900 mb-4 leading-tight">
-                What teams love about<br />
-                Marketing Simplified?
-              </h2>
-              <p className="text-base text-gray-700 mb-6 leading-relaxed">
-                From setup to reporting, Marketing Simplified helps teams stay aligned, efficient, and confident in every campaign.
-              </p>
-            </div>
-
-            <div className="overflow-x-auto scrollbar-hide snap-x snap-mandatory scroll-smooth mb-2 pb-4 -mx-14 px-10">
-              <div className="flex gap-4 min-w-max">
-                {/* Testimonial 1 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 min-w-[280px] flex-shrink-0 snap-center flex flex-col">
-                  <p className="text-gray-700 mb-4 italic flex-grow relative text-center text-sm">
-                    <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-4 w-3 h-3" />
-                    <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-4 w-3 h-3" />
-                    Finally, one place to manage<br />
-                    budgets, assets, and approvals.<br />
-                    We cut our setup time by half.
-                  </p>
-                  <div className="flex items-center gap-3 mt-auto">
-                    <div className="w-10 h-10 bg-blue-100 rounded-full overflow-hidden flex-shrink-0">
-                      <img
-                        src="https://i.pravatar.cc/150?img=68"
-                        alt="Jack Wilson"
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <div>
-                      <div className="font-bold text-gray-900 text-sm">Jack Wilson</div>
-                      <div className="text-xs text-gray-600">Senior Media Buyer</div>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Testimonial 2 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 min-w-[280px] flex-shrink-0 snap-center flex flex-col">
-                  <p className="text-gray-700 mb-4 italic flex-grow relative text-center text-sm">
-                    <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-4 w-3 h-3" />
-                    <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-4 w-3 h-3" />
-                    I love how the review and<br />
-                    approval flow feels natural.<br />
-                    No more endless Slack threads.
-                  </p>
-                  <div className="flex items-center gap-3 mt-auto">
-                    <div className="w-10 h-10 bg-green-100 rounded-full overflow-hidden flex-shrink-0">
-                      <img
-                        src="https://i.pravatar.cc/150?img=47"
-                        alt="Eve Turner"
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <div>
-                      <div className="font-bold text-gray-900 text-sm">Eve Turner</div>
-                      <div className="text-xs text-gray-600">Designer</div>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Testimonial 3 */}
-                <div className="bg-white rounded-xl shadow-lg p-6 min-w-[280px] flex-shrink-0 snap-center flex flex-col">
-                  <p className="text-gray-700 mb-4 italic flex-grow relative text-center text-sm">
-                    <img src="/opening quote.png" alt="Opening quote" className="absolute top-0 -left-4 w-3 h-3" />
-                    <img src="/closing quote.png" alt="Closing quote" className="absolute bottom-0 -right-4 w-3 h-3" />
-                    The reporting dashboard gives<br />
-                    instant clarity. We spot<br />
-                    underperforming channels in<br />
-                    minutes.
-                  </p>
-                  <div className="flex items-center gap-3 mt-auto">
-                    <div className="w-10 h-10 bg-orange-100 rounded-full overflow-hidden flex-shrink-0">
-                      <img
-                        src="https://i.pravatar.cc/150?img=59"
-                        alt="Grace Lee"
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <div>
-                      <div className="font-bold text-gray-900 text-sm">Grace Lee</div>
-                      <div className="text-xs text-gray-600">Data Analyst</div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            <div className="text-center">
-              <button className="flex items-center px-6 py-3 bg-blue-800 text-white rounded-full hover:bg-blue-900 transition text-center font-medium text-sm gap-4 shadow-lg">
-                See All Reviews
-                <ArrowRight className="w-4 h-4" />
-              </button>
-            </div>
-          </div>
-        </div>
-      </section>
-    </>
-  );
+      <motion.div
+        initial={{ opacity: 0, y: 20 }}
+        animate={isInView ? { opacity: 1, y: 0 } : {}}
+        transition={{ duration: 0.6, delay: 0.4 }}
+        className="text-center"
+      >
+        <button className="inline-flex items-center gap-2 px-8 py-4 bg-brand-gradient text-white rounded-full hover:saturate-150 transition-all font-medium text-lg glow-brand">
+          See All Reviews
+          <ArrowRight className="w-5 h-5" />
+        </button>
+      </motion.div>
+    </section>
+  )
 }
